fixed datatype isssues

+++ src/main/java/com/example/ABCBankDetails/dto/PayeeAddRequestDTO.java
@@ -12,13 +12,13 @@ public class PayeeAddRequestDTO implements Serializable {
 	private int payeeId;
 	
 	
-	public Long getCustomerId() {
+	public int getCustomerId() {
 		return customerId;
 	}
 	public void setCustomerId(int customerId) {
 		this.customerId = customerId;
 	}
-	public Long getPayeeId() {
+	public int getPayeeId() {
 		return payeeId;
 	}
 	public void setPayeeId(int payeeId) {
